+++ firmware/uNode_2/websocket.cpp
@@ -221,6 +221,10 @@ void broadcastStatus() {
 
   doc["liveProtocol"] =
     config.liveProtocol;
+  doc["sacnSourceName"] =
+    config.sacnSourceName;
+  doc["sacnPriority"] =
+    config.sacnPriority;
 
   doc["sacnUniverse"] =
     getSacnUniverse();